+++ src/entries/wvui-icons.ts
@@ -1,0 +1,6 @@
+// Export version as a named export so that the default export so that the default export can be
+// passed to the Vue app instance's components directly.
+export const version = VERSION;
+
+// Export all icons available in the library.
+export * as icons from '@/themes/icons';
